FaaSMem-core/requirements.txt
flask
gevent
requests
docker
PyYAML
numpy
pandas
tqdm
seaborn
matplotlib
